web/requirements.txt
Flask==3.1.2
requests==2.31.0
psycopg2-binary
paho-mqtt==1.6.1
cryptography==41.0.5
gunicorn
gevent>=23.9.1